+++ Src/DependencyCollector/Shared/Implementation/RemoteDependencyConstants.cs
@@ -10,6 +10,7 @@
         public const string AzureTable = "Azure table";
         public const string AzureQueue = "Azure queue";
         public const string AzureDocumentDb = "Azure DocumentDB";
+        public const string AzureEventHubs = "Azure Event Hubs";
         public const string AzureServiceBus = "Azure Service Bus";
         public const string AzureIotHub = "Azure IoT Hub";